cmd/compile/internal/ssa: join return blocks
For functions with one or more return values,
fix up the CFG so that there is only one return
block. This reduces the size of the function by
the number of stack slots consumed by return params
times the number of original return blocks.

Functions with odd return behavior, like functions
with named return values that use naked returns, are
silently ignored.

This change reduces the size of the Go tool by about 1%,
from 10.5MB to 10.4MB (on darwin/amd64).

Change-Id: I16db53afd6b9910e4338ad02c6ad88f0c6d70cbd